Add demo spring boot app that uses OAuth 2.0 to connect to RabbitMQ
It uses Spring Oauth 2.0 library to interact with Authorization server
in order to get an access token using client Grant type flow.
It does not leverage all the auto-configuration in the project because
it is meant for applications which receives user requests via http
(servlet, reactor) and accesses a resource server on behalf of the end
user (a.k.a. resource owner). This is not our case hence we do not use
the auuth annotations.