Take OAuth client credentials from the Cloud environment
Before this change, we were hard-coding it because we did not have a way
to pass the oauth info via the AmqpServiceInfo.
Now we have created an extended version called AmqpOAuthServiceInfo
bound to the service label "rabbitmq-oauth" which reads the client_id,
client_secret and auth_domain from within the credentials json element.

The next step is to fully integrate the AmqpOAuthServiceInfoCreator into
the ServiceConnectors so that the application does not need to create
them by hand.